+++ app/models/school_workforce_census.rb
@@ -1,0 +1,3 @@
+class SchoolWorkforceCensus < ApplicationRecord
+  self.table_name = "school_workforce_censuses"
+end