+++ config/routes.rb
@@ -96,6 +96,7 @@ Rails.application.routes.draw do
     end
 
     resources :qualification_report_uploads, only: [:new, :create]
+    resources :school_workforce_census_data_uploads, only: [:new, :create]
 
     resources :payroll_runs, only: [:index, :new, :create, :show] do
       resources :payment_confirmation_report_uploads, only: [:new, :create]